Create a new user data table
To allow the data to be anonymised, it is easier to store all of the necessary data in a separate table. While this goes against normal principles, this was the best decision for this study.

The user data table stores a link to the user id table, information about the pre-study questionnaire, information about the study, information about the post study questionnaire and about the three knowledge retention quizzes.

[GG-65]